#189 refactor: StockPage AsyncBoundary 적용 및 UI 수정
