+++ src/api/stock/queries/useStockPageQeury.ts
@@ -1,11 +1,12 @@
-import { useQuery } from "@tanstack/react-query";
+import { useSuspenseQuery } from "@tanstack/react-query";
 import { getStockPage } from "..";
 import { stockKeys } from "./queryKeys";
 
 export default function useStockPageQuery(tickerSymbol: string) {
-  return useQuery({
+  return useSuspenseQuery({
     queryKey: stockKeys.page(tickerSymbol).queryKey,
     queryFn: () => getStockPage(tickerSymbol),
+    retry: false,
     select: (res) => res.data,
   });
 }
